Template: vite-cf-DO-runner
Base template for I would like you to create me a simple image hosting website.  Must be able to pick and retain a directory folder on load.  User should not have to interact with it, site is hosted privately so no need to username or password.  
1. Enter  URL
2. Site loads and displays images
3. site must be organized and can easily view thumbnails, exif data, like google photos or onedrive photos.

Go!
Can be in  react, since I already have bun installed.